+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Adds a `Goatee.Types()` class for clarity. Types can be used to specify element types for example in `<input>` elements.
+- Adds a "Goatee.Types()" class for clarity. Types can be used to specify element types for example in "<input>" elements.
 - Adds support for the "<div"> element using the "Goatee.Container()" class.
 - Adds support for the "<section>" element using the "Goatee.Section()" class.
 - Adds support for the "<main>" element using the "Goatee.Main()" class.